What: remove support of go1.6 and older
Why:

  * upgrade the min go version supported

How:

  * nop
